Restrict the search to bugs with a given status (--status flag)
Allow the query to Launchpad to be restricted to bugs with a specified
status. The implementation relies on the Launchpad API treating a query
for bugs with an empty status as a query for which no status has been
specified, thus returning bugs with any status. This behavior is not
explicitly documented in the Launchpad API [0].

[0] https://launchpad.net/+apidoc/1.0.html