CLI: Add --urls to print only URLs (#17)
The current system of launching a browser with a long list of bugs is a
poor experience. As proposed, this adds the --urls option which will
only print the urls of the bugs to triage. Then a user can pass that
output (e.g `--urls | xargs webbrowser`) and avoid the issue.

Fixes #16