feat: enable filtering bugs by tags
Update argument parser with --any-tag to switch tag combinator.
Add a --tag argument which can search by tag: -tfoo -tbar
and which can do negative search --tag=-notfoo -t-notbar.

The dash prefix is a bit fiddly with argparse's parser, so I left an
example in the help string to help around that.